+++ upcoming-release-notes/5837.md
@@ -1,0 +1,6 @@
+---
+category: Maintenance
+authors: [matt-fidd]
+---
+
+Refactor rules framework
